Add files via upload
Changes done: now there is a a* algorithm generating the path for the arm to follow, works well. Made some attributes private as they weren't needed as public, might even do that on the node and grid scripts.  The evolution is script is only enabled after the path is found by the a* algorithm, in the pathfhinding alogirthm made the seeker always be changed to the origin upon starting. Next steps: i really suspect the gizmos is making this thing really hard on the CPU, might stop drawing on gizmos entirely and rather creating gameobjects, but its future work