Tons of New Features - Obstacles and new Fitness
Random obstacles are now generated in the createScene script, and get their obstacle layers such a layer already was possible at the grid and pathfinding scripts and thus were utilized. The obstacles are added to a gameobject list used by the evolution script, and the minimum distance of any joint to any obstacle is calculated, this can be optimized just as in the 2d version using a weighted grid, that will be done later. a new fitness function considering and capping the two things, distance to the objective and distance to obstacle is now used.